overhaul with CI/CD, envvar-renames and lots of fixes to make CI happy
