+++ CONTRIBUTORS.rst
@@ -1,5 +1,8 @@
 Contributors
 ============
 
-- Jens W. Klein, jk@kleinundpartner.at
-- Katja Süss, Rohberg, @ksuess
+- Peter Holzer - Initiative, idea and testing.
+- Jens W. Klein, jk@kleinundpartner.at - Concept & code.
+- Katja Süss, Rohberg, @ksuess - Text analysis code and configuration.
+
+Initial implementation was made possible by `Evangelisch-reformierte Landeskirche des Kantons Zürich <https://zhref.ch/>`_.
