--- constraints.txt
@@ -1,15 +1,0 @@
-amqp==2.6.1
-billiard==3.6.4.0
-CacheControl==0.12.6
-celery==4.4.7
-certifi==2021.5.30
-chardet==4.0.0
-elasticsearch==7.13.1
-idna==2.10
-kombu==4.6.11
-msgpack==1.0.2
-pytz==2021.1
-redis==3.5.3
-requests==2.25.1
-urllib3==1.26.5
-wcwidth==0.2.5
