+++ examples/.env
@@ -5,11 +5,11 @@
 # Then `source .env` and start the ingest-service with: celery -A collective.elastic.ingest.celery.app worker -l debug
 export CELERY_BROKER=redis://localhost:6379/0
 
-export ELASTICSEARCH_INGEST_SERVER=localhost:9200
-export ELASTICSEARCH_INGEST_USE_SSL=1
-export OPENSEARCH=1
-export ELASTICSEARCH_INGEST_LOGIN=admin
-export ELASTICSEARCH_INGEST_PASSWORD=admin
+export INGEST_SERVER=localhost:9200
+export INGEST_USE_SSL=1
+export INGEST_OPENSEARCH=1
+export INGEST_LOGIN=admin
+export INGEST_PASSWORD=admin
 
 export PLONE_SERVICE=http://localhost:8080
 export PLONE_PATH=Plone
