+++ mypy.ini
@@ -1,0 +1,2 @@
+[mypy]
+mypy_path=./src
