--- setup.cfg
@@ -1,11 +1,0 @@
-[mypy]
-ignore_missing_imports = True
-
-[bdist_wheel]
-universal=1
-
-[isort]
-profile = plone
-
-[zest.releaser]
-create-wheel = yes
