--- src/collective/elastic/ingest/analysis/__init__.py
@@ -1,1 +1,0 @@
-from .analysis import update_analysis
