+++ tox.ini
@@ -1,0 +1,53 @@
+# Generated from:
+# https://github.com/plone/meta/tree/master/config/default
+# See the inline comments on how to expand/tweak this configuration file
+[tox]
+# We need 4.4.0 for constrain_package_deps.
+min_version = 4.4.0
+envlist =
+    py38
+    py39
+    py310
+    py311
+    py312
+    test
+    lint
+
+[gh-actions]
+python =
+    3.8: py38
+    3.9: py39
+    3.10: py310
+    3.11: py311
+    3.12: py312, lint
+
+[testenv]
+use_develop = true
+skip_install = false
+constrain_package_deps = true
+
+commands =
+    pytest {posargs}
+extras =
+    test
+    elasticsearch
+    opensearch-py
+
+[testenv:format]
+description = automatically reformats code
+skip_install = true
+deps =
+    pre-commit
+commands =
+    pre-commit run -a pyupgrade
+    pre-commit run -a isort
+    pre-commit run -a black
+
+[testenv:lint]
+description = run linters that will help improve the code style
+skip_install = true
+deps =
+    pre-commit
+commands =
+    pre-commit run -a
+
